WIP -- Cannot make progress, because metadata.name may be undefined when AWSCluster is seen by defaulting webhook
I learned that a mutating webhook cannot assume that the metadata.name
field is defined. For example, if the AWSCluster object has
metadata.generateName defined, and metadata.name undefined, the
defaulting webhook (a mutating webhook) cannot generate any values from
the cluster name.
https://github.com/kubernetes/kubernetes/issues/76680#issuecomment-484110717

This means it's not possible to default the ELB name, because it is
generated from the cluster name, which is not defined until after the
object has passed through all the webhooks and is stored.